+++ src/Controller/CRUD/QuizCRUDController.php
@@ -1,41 +1,57 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\CRUD;
 
 use App\Entity\Quiz;
 use App\Form\QuizType;
+use App\Repository\AnswerRepository;
+use App\Repository\QuestionRepository;
 use App\Repository\QuizRepository;
+use App\Representation\RepresentAs;
+use App\Representation\RepresentationType;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Tools\Pagination\Paginator;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\FormInterface;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route("/app/quiz")]
-class QuizCRUDController extends AbstractController
+class QuizCRUDController extends CRUDController
 {
-    public function __construct(protected EntityManagerInterface $entityManager, protected QuizRepository $quizRepository)
+    public function __construct(
+        protected QuizRepository         $quizRepository,
+        protected QuestionRepository     $questionRepository,
+        protected AnswerRepository       $answerRepository,
+        protected EntityManagerInterface $entityManager,
+        protected Security               $security,
+    )
     {
+        parent::__construct($this->entityManager);
     }
 
     #[Route("/create", name: "quiz_create", methods: ['POST', 'GET'])]
     #[Route("/{quiz}/edit", name: "quiz_edit", methods: ['POST', 'GET'])]
     public function createQuiz(Request $request, Quiz $quiz = null): Response
     {
-
         $quiz ??= (new Quiz())
             ->setVersion(1);
 
         $form = $this->createForm(QuizType::class, $quiz, [
+            'action' => $request->getRequestUri(),
             'attr' => [
                 'data-turbo-action' => 'advance'
             ]
         ]);
 
         if ($this->handleForm($form, $request)) {
-            return $this->redirectToRoute('quiz_list', ['search' => $quiz->getValue()]);
+            if (!$this->currentRouteIs($request, 'quiz_create')) {
+                $this->addFlash('success', "Quiz '{$quiz->getValue()}' was created.");
+            } else {
+                $this->addFlash('success', "Quiz '{$quiz->getValue()}' was updated.");
+            }
+
+            return $this->redirectToRoute('go_through_quiz', ['quiz' => $quiz->getId()]);
         }
 
         return $this->render(
@@ -47,7 +63,9 @@ class QuizCRUDController extends AbstractController
     }
 
     #[Route("/list", name: "quiz_list", methods: ['GET'])]
-    public function listQuizzes(Request $request): Response
+    #[RepresentAs(RepresentationType::TURBO, template: '/CRUD/quiz/frame/_list-quiz.html.twig')]
+    #[RepresentAs(RepresentationType::HTML, template: '/CRUD/quiz/list.html.twig')]
+    public function listQuizzes(Request $request): array
     {
         $perPage = max($request->query->getInt('perPage', 1), 10);
         $page = max($request->query->getInt('page', 1), 1);
@@ -60,7 +78,7 @@ class QuizCRUDController extends AbstractController
             ->setMaxResults($perPage)
             ->setFirstResult(($page - 1) * $perPage);
 
-        if(!empty($search)) {
+        if (!empty($search)) {
             $queryBuilder
                 ->where('quiz.value LIKE :search')
                 ->setParameter('search', "%{$search}%");
@@ -68,14 +86,12 @@ class QuizCRUDController extends AbstractController
 
         $paginator = new Paginator($queryBuilder);
 
-        return $this->render(
-            'CRUD/quiz/list.html.twig',
-            [
-                'list' => $paginator->getIterator(),
-                'totalPages' => ceil($paginator->count() / $perPage),
-                'perPage' => $perPage,
-                'page' => $page,
-            ]);
+        return [
+            'list' => $paginator->getIterator(),
+            'totalPages' => ceil($paginator->count() / $perPage),
+            'perPage' => $perPage,
+            'page' => $page,
+        ];
     }
 
     #[Route("/{quiz}/delete", name: "quiz_delete", methods: ['DELETE'])]
@@ -90,21 +106,4 @@ class QuizCRUDController extends AbstractController
 
         return $this->redirectToRoute('quiz_list', status: Response::HTTP_SEE_OTHER);
     }
-
-    public function handleForm(FormInterface $form, Request $request): bool
-    {
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            /** @var Quiz $quiz */
-            $this->entityManager->persist($quiz = $form->getData());
-            $this->entityManager->flush();
-            $this->addFlash('success', "Quiz '{$quiz->getValue()}' was created.");
-
-            return true;
-        }
-
-        return false;
-    }
-
 }