+++ src/Representation/RepresentationType.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Representation;
+
+enum RepresentationType: string
+{
+    case TURBO = 'turbo';
+    case HTML = 'html';
+    case FORM_SUBMITTED = 'form_submitted';
+    case REDIRECT = 'redirect';
+}