+++ CHANGELOG.md
@@ -5,11 +5,17 @@
 **Added**
 - Create a home page with navigation.
 - QR-1: Add possibility to create/edit/delete quizzes.
+- QR-13: Added possibility to add questions.
+- QR-22: Added possibility to add answers.
+- QR-8: Turbo frames fragmented rendering.
 
 **Changed**
+- QR-17: Replace alerts with toasts
 
 **Fixed**
 - Change `Submit` button to `Update` button when editing an answer.
+- QR-24: Quiz search is unfocused when page is reloaded
+- QR-6: Change the URL's to reduce the amount of variables that are required for each request
 
 **Requirements**